matcher: fix fp with missing select key and NOT
Event has two outputs - value and indication on weather or not value can
be even extracted (key exists in message). But matcher was built to only
return one boolean. That boolean can be flipped with Not struct, meaning
that False returned by missing key can be flipped and rule can evaluate
as true. Thus, some rules would always match, even when not applicable.