Realized after finishing my fix to properly translate escapes from Sigma to Glob that the glob library also has special meaning for brackets (square and curly).
With this change I implemented a specialzed QuoteMeta-type function into
the escape function, where as we encounter any of the square or curly
brackets, we escape them automatically (these brackets don't have a
special meaning in Sigma rules today to my knowledge).

This could become more complicated if Sigma ever decides to do something
with '[', ']' or '{', '}'.